+++ jest.config.js
@@ -1,16 +1,18 @@
 module.exports = {
   transform: {
-    '^.+\\.tsx?$': ['ts-jest']
+    '^.+\\.tsx?$': 'ts-jest',
+    '^.+\\.html?$': './build/jest.html-loader.js'
   },
   testEnvironment: 'jsdom',
   roots: ['src/modules', 'src/polyfills'],
   testRegex: '/test/(.+)\\.test\\.ts$',
-  moduleFileExtensions: ['ts', 'js', 'json'],
+  moduleFileExtensions: ['ts', 'js', 'json', 'html'],
   coverageDirectory: '.report',
   coverageReporters: ['lcov', 'html'],
   setupFiles: [
     './src/modules/esl-utils/test/deviceDetector.mock.ts',
-    './src/modules/esl-utils/test/matchMedia.mock.ts'
+    './src/modules/esl-utils/test/matchMedia.mock.ts',
+    './src/modules/esl-utils/test/resizeObserver.mock.ts'
   ],
   collectCoverageFrom: [
     'src/modules/**/*.ts',
@@ -19,8 +21,8 @@ module.exports = {
     // cumulative exclude
     '!src/modules/*.ts',
     '!src/modules/*/*.ts',
-    // beta modules exclude
-    '!src/modules/beta/**',
+    // draft modules exclude
+    '!src/modules/draft/**',
     // libs exclude
     '!**/node_modules/**'
   ]