+++ src/modules/esl-footnotes/core/esl-note.ts
@@ -16,7 +16,7 @@ import type {IMediaQueryCondition} from '../../esl-media-query/core/conditions/m
 
 @ExportNs('Note')
 export class ESLNote extends ESLBaseElement {
-  public static is = 'esl-note';
+  public static override is = 'esl-note';
   public static observedAttributes = ['tooltip-shown', 'ignore'];
 
   /** Timeout before activating note (to have time to show content with this note) */
@@ -95,14 +95,14 @@ export class ESLNote extends ESLBaseElement {
   }
 
   @ready
-  protected connectedCallback(): void {
+  protected override connectedCallback(): void {
     this.init();
     super.connectedCallback();
     this._sendResponseToFootnote();
   }
 
   @ready
-  protected disconnectedCallback(): void {
+  protected override disconnectedCallback(): void {
     super.disconnectedCallback();
     this._$footnotes?.unlinkNote(this);
     this.restore();
@@ -114,15 +114,20 @@ export class ESLNote extends ESLBaseElement {
       this._$footnotes?.turnOffHighlight(this);
     }
     if (attrName === 'ignore') {
-      memoize.clear(this, 'queryToIgnore');
-      this.$$on(this._onBPChange);
+      this.updateIgnoredQuery();
     }
+  }
+
+  /** Revise the settings for ignoring the note */
+  public updateIgnoredQuery(): void {
+    memoize.clear(this, 'queryToIgnore');
+    this.$$on(this._onBPChange);
+    this._onBPChange();
   }
 
   /** Gets attribute value from the closest element with group behavior settings */
   protected getClosestRelatedAttr(attrName: string): string | null {
-    const tagName = (this.constructor as typeof ESLBaseElement).is;
-    const relatedAttrName = `${tagName}-${attrName}`;
+    const relatedAttrName = `${this.baseTagName}-${attrName}`;
     const $closest = this.closest(`[${relatedAttrName}]`);
     return $closest ? $closest.getAttribute(relatedAttrName) : null;
   }