+++ src/modules/esl-mixin-element/ui/esl-mixin-element.ts
@@ -4,7 +4,6 @@ import {ESLEventUtils} from '../../esl-utils/dom/events';
 import {CSSClassUtils} from '../../esl-utils/dom/class';
 import {ESLMixinRegistry} from './esl-mixin-registry';
 
-import type {AttributeTarget} from '../../esl-utils/dom/attr';
 import type {
   ESLEventListener,
   ESLListenerCriteria,
@@ -12,13 +11,14 @@ import type {
   ESLListenerEventMap,
   ESLListenerHandler
 } from '../../esl-utils/dom/events';
+import type {ESLDomElementRelated} from '../../esl-utils/abstract/dom-target';
 
 /**
  * Base class for mixin elements.
  * Mixin elements attaches to the DOM element via attribute.
  * Allows multiple mixin elements per one DOM element
  */
-export class ESLMixinElement implements AttributeTarget {
+export class ESLMixinElement implements ESLDomElementRelated {
   /** Root attribute to identify mixin targets. Should contain dash in the name. */
   static is: string;
   /** Additional observed attributes */
@@ -36,13 +36,17 @@ export class ESLMixinElement implements AttributeTarget {
 
   /** Callback of mixin instance initialization */
   public connectedCallback(): void {
-    const constructor = this.constructor as typeof ESLMixinElement;
-    if (constructor.observedAttributes.length) {
+    const {observedAttributes} = this.constructor as typeof ESLMixinElement;
+    if (observedAttributes.length) {
       this._attr$$ = new MutationObserver(this._onAttrMutation.bind(this));
-      this._attr$$.observe(this.$host, {attributes: true, attributeFilter: constructor.observedAttributes});
+      this._attr$$.observe(this.$host, {
+        attributes: true,
+        attributeFilter: observedAttributes,
+        attributeOldValue: true
+      });
     }
 
-    ESLEventUtils.descriptors(this).forEach((desc) => ESLEventUtils.subscribe(this, desc));
+    ESLEventUtils.subscribe(this);
   }
 
   /** Callback to execute on mixin instance destroy */
@@ -52,7 +56,10 @@ export class ESLMixinElement implements AttributeTarget {
     ESLEventUtils.unsubscribe(this);
   }
 
-  /** Callback to handle changing of additional attributes */
+  /**
+   * Callback to handle changing of additional attributes.
+   * Happens when attribute accessed for writing independently of the actual value change
+   */
   public attributeChangedCallback(name: string, oldValue: string | null, newValue: string | null): void {}
 
   /** Attribute change mutation record processor */