+++ src/modules/esl-scrollbar/core/esl-scrollbar.ts
@@ -5,7 +5,7 @@ import {rafDecorator} from '../../esl-utils/async/raf';
 import {isMouseEvent, isTouchEvent, getTouchPoint, getOffsetPoint} from '../../esl-utils/dom/events';
 import {isRelativeNode} from '../../esl-utils/dom/traversing';
 import {ESLTraversingQuery} from '../../esl-traversing-query/core';
-import {RTLUtils} from '../../esl-utils/dom/rtl';
+import {isRTL, RTLScroll, normalizeScrollLeft} from '../../esl-utils/dom/rtl';
 
 /**
  * ESLScrollbar is a reusable web component that replaces the browser's default scrollbar with
@@ -15,7 +15,7 @@ import {RTLUtils} from '../../esl-utils/dom/rtl';
  */
 @ExportNs('Scrollbar')
 export class ESLScrollbar extends ESLBaseElement {
-  public static is = 'esl-scrollbar';
+  public static override is = 'esl-scrollbar';
   public static observedAttributes = ['target', 'horizontal'];
 
   /** Horizontal scroll orientation marker */
@@ -55,20 +55,20 @@ export class ESLScrollbar extends ESLBaseElement {
   protected _mutationObserver = new MutationObserver((rec) => this.updateContentObserve(rec));
 
   @ready
-  protected connectedCallback(): void {
+  protected override connectedCallback(): void {
     super.connectedCallback();
-    this.findTarget();
     this.render();
+    this.findTarget();
   }
 
   @ready
-  protected disconnectedCallback(): void {
+  protected override disconnectedCallback(): void {
     this.unbindTargetEvents();
     this._scrollTimer && window.clearTimeout(this._scrollTimer);
   }
 
   protected attributeChangedCallback(attrName: string, oldVal: string, newVal: string): void {
-    if (!this.connected && oldVal === newVal) return;
+    if (!this.connected || oldVal === newVal) return;
     if (attrName === 'target') this.findTarget();
     if (attrName === 'horizontal') this.refresh();
   }
@@ -104,16 +104,19 @@ export class ESLScrollbar extends ESLBaseElement {
 
   protected bindTargetEvents(): void {
     if (!this.$target) return;
+    // Container resize/scroll observers/listeners
     if (document.documentElement === this.$target) {
-      this.$$on(this._onScrollOrResize);
+      this.$$on({event: 'scroll resize', target: window}, this._onScrollOrResize);
     } else {
       this.$$on({event: 'scroll', target: this.$target}, this._onScrollOrResize);
       this._resizeObserver.observe(this.$target);
-      this._mutationObserver.observe(this.$target, {childList: true});
-      Array.from(this.$target.children).forEach((el) => this._resizeObserver.observe(el));
     }
+    // Subscribes to the child elements resizes
+    this._mutationObserver.observe(this.$target, {childList: true});
+    Array.from(this.$target.children).forEach((el) => this._resizeObserver.observe(el));
   }
 
+  /** Resubscribes resize observer on child elements when container content changes */
   protected updateContentObserve(recs: MutationRecord[] = []): void {
     if (!this.$target) return;
     const contentChanges = recs.filter((rec) => rec.type === 'childList');
@@ -167,7 +170,7 @@ export class ESLScrollbar extends ESLBaseElement {
   /** Relative position value (between 0.0 and 1.0) */
   public get position(): number {
     if (!this.$target) return 0;
-    const scrollOffset = this.horizontal ? RTLUtils.normalizeScrollLeft(this.$target) : this.$target.scrollTop;
+    const scrollOffset = this.horizontal ? normalizeScrollLeft(this.$target) : this.$target.scrollTop;
     return this.scrollableSize ? (scrollOffset / this.scrollableSize) : 0;
   }
 
@@ -179,8 +182,8 @@ export class ESLScrollbar extends ESLBaseElement {
   /** Normalizes position value (between 0.0 and 1.0) */
   protected normalizePosition(position: number): number {
     const relativePosition = Math.min(1, Math.max(0, position));
-    if (!RTLUtils.isRtl(this.$target) || !this.horizontal) return relativePosition;
-    return RTLUtils.scrollType === 'negative' ? (relativePosition - 1) : (1 - relativePosition);
+    if (!isRTL(this.$target) || !this.horizontal) return relativePosition;
+    return RTLScroll.type === 'negative' ? (relativePosition - 1) : (1 - relativePosition);
   }
 
   /** Scrolls target element to passed position */
@@ -334,7 +337,6 @@ export class ESLScrollbar extends ESLBaseElement {
    * Handler for scroll and resize events
    * @param event - instance of 'resize' or 'scroll' event
    */
-  @listen({event: 'resize scroll', passive: true, auto: false})
   protected _onScrollOrResize(event: Event): void {
     if (event.type === 'scroll' && this.dragging) return;
     this._deferredRefresh();