+++ src/modules/esl-tooltip/core/esl-tooltip.ts
@@ -19,16 +19,16 @@ export interface TooltipActionParams extends PopupActionParams {
 
 @ExportNs('Tooltip')
 export class ESLTooltip extends ESLPopup {
-  static is = 'esl-tooltip';
+  static override is = 'esl-tooltip';
 
   /**
    * Tooltip position relative to the trigger.
    * Currently supported: 'top', 'bottom', 'left', 'right' position types ('top' by default)
    */
-  @attr({defaultValue: 'top'}) public position: PositionType;
+  @attr({defaultValue: 'top'}) public override position: PositionType;
 
   /** Tooltip behavior if it does not fit in the window ('fit' by default) */
-  @attr({defaultValue: 'fit'}) public behavior: string;
+  @attr({defaultValue: 'fit'}) public override behavior: string;
 
   /** Disable arrow at Tooltip */
   @boolAttr() public disableArrow: boolean;
@@ -71,7 +71,7 @@ export class ESLTooltip extends ESLPopup {
     this.sharedInstance.hide(params);
   }
 
-  public connectedCallback(): void {
+  public override connectedCallback(): void {
     if (!this.disableArrow) {
       this._appendArrow();
     }
@@ -81,7 +81,7 @@ export class ESLTooltip extends ESLPopup {
   }
 
   /** Sets initial state of the Tooltip */
-  protected setInitialState(): void {}
+  protected override setInitialState(): void {}
 
   /** Creates arrow at Tooltip */
   @memoize()
@@ -98,7 +98,7 @@ export class ESLTooltip extends ESLPopup {
   }
 
   /** Actions to execute on show Tooltip. */
-  public onShow(params: TooltipActionParams): void {
+  public override onShow(params: TooltipActionParams): void {
     if (params.disableArrow) {
       this.disableArrow = params.disableArrow;
     }
@@ -117,7 +117,7 @@ export class ESLTooltip extends ESLPopup {
   }
 
   /** Actions to execute on Tooltip hiding. */
-  public onHide(params: TooltipActionParams): void {
+  public override onHide(params: TooltipActionParams): void {
     this._updateActivatorState(false);
     super.onHide(params);
     document.body.removeChild(this);
@@ -129,7 +129,7 @@ export class ESLTooltip extends ESLPopup {
   /**
    * Actions to execute after showing of popup.
    */
-  protected afterOnShow(): void {
+  protected override afterOnShow(): void {
     super.afterOnShow();
     this.focus({preventScroll: true});
   }
@@ -137,12 +137,12 @@ export class ESLTooltip extends ESLPopup {
   /**
    * Actions to execute before hiding of popup.
    */
-  protected beforeOnHide(): void {
+  protected override beforeOnHide(): void {
     this.activator?.focus({preventScroll: true});
   }
 
   @listen({inherit: true})
-  protected _onKeyboardEvent(e: KeyboardEvent): void {
+  protected override _onKeyboardEvent(e: KeyboardEvent): void {
     super._onKeyboardEvent(e);
     if (e.key === TAB) this._onTabKey(e);
   }