+++ src/modules/esl-utils/dom/events/target.ts
@@ -1,9 +1,14 @@
+import {overrideEvent} from './misc';
+
 /**
  * Synthetic implementation of EventTarget
  * Replicates behavior of native event
  * Doesn't give explicit access to callback storage
  */
 export class SyntheticEventTarget implements EventTarget {
+  // Event type to use in the shortcutted calls
+  public static DEFAULT_EVENT = 'change';
+
   private readonly _listeners: Record<string, EventListenerOrEventListenerObject[]> = {};
 
   public hasEventListener(): boolean;
@@ -17,7 +22,7 @@ export class SyntheticEventTarget implements EventTarget {
   public addEventListener(callback: EventListenerOrEventListenerObject): void;
   public addEventListener(type: string, callback: EventListenerOrEventListenerObject): void;
   public addEventListener(type: string | EventListenerOrEventListenerObject, callback?: EventListenerOrEventListenerObject): void {
-    if (typeof type !== 'string') return this.addEventListener('change', type);
+    if (typeof type !== 'string') return this.addEventListener((this.constructor as typeof SyntheticEventTarget).DEFAULT_EVENT, type);
 
     validateEventListenerType(callback);
     if (this._listeners[type] && this._listeners[type].includes(callback!)) return;
@@ -28,18 +33,21 @@ export class SyntheticEventTarget implements EventTarget {
   public removeEventListener(callback: EventListenerOrEventListenerObject): void;
   public removeEventListener(type: string, callback: EventListenerOrEventListenerObject): void;
   public removeEventListener(type: string | EventListenerOrEventListenerObject, callback?: EventListenerOrEventListenerObject): void {
-    if (typeof type !== 'string') return this.removeEventListener('change', type);
+    if (typeof type !== 'string') return this.removeEventListener((this.constructor as typeof SyntheticEventTarget).DEFAULT_EVENT, type);
 
     validateEventListenerType(callback);
     if (!this._listeners[type]) return;
     this._listeners[type] = this._listeners[type].filter((cb) => cb !== callback);
   }
 
-  public dispatchEvent(e: Event): boolean {
-    const target = (): EventTarget => this; // use get due IE specific
-    Object.defineProperty(e, 'target', {get: target, enumerable: true});
-    Object.defineProperty(e, 'currentTarget', {get: target, enumerable: true});
-    Object.defineProperty(e, 'srcElement', {get: target, enumerable: true});
+  public dispatchEvent(e: Event): boolean;
+  /** @deprecated use `overrideEvent` to declare `target` */
+  public dispatchEvent(e: Event, target?: EventTarget): boolean;
+  public dispatchEvent(e: Event, target?: EventTarget): boolean {
+    overrideEvent(e, 'currentTarget', this);
+    if (target) overrideEvent(e, 'target', target); // TODO: remove in 5.0.0
+    if (!e.target) overrideEvent(e, 'target', this);
+    if (!e.srcElement) overrideEvent(e, 'srcElement', e.target); // TODO: remove in 5.0.0
     this._listeners[e.type]?.forEach((listener) => {
       if (typeof listener === 'function') listener.call(this, e);
       else listener.handleEvent.call(listener, e);